Modify index.html and stylesheets
Added div element to index.html for potential navigation bar in the
future.

Modified values in styles-index.css to make anchor element links to
align with each other better. Moved font-family to cover the body
of the document. Added a dashed border instead of none to anchor
mousehover.

Pasted code for HTML's inherited values for all stylesheets as well as
added the same font-family to cover all styles for consistency across
all pages.